refactor(vize): split check/tsconfig_inputs into cohesive modules
