+++ crates/vize/src/commands/check/tsconfig_inputs/jsonc.rs
@@ -1,0 +1,119 @@
+//! Minimal JSONC support: strip comments and trailing commas before parsing.
+
+use serde_json::Value;
+
+pub(crate) fn parse_jsonc_value(content: &str) -> Result<Value, serde_json::Error> {
+    let stripped = strip_json_comments(content);
+    let normalized = strip_trailing_commas(&stripped);
+    serde_json::from_str(&normalized)
+}
+
+fn strip_json_comments(content: &str) -> std::string::String {
+    let mut output = std::string::String::with_capacity(content.len());
+    let mut chars = content.chars().peekable();
+    let mut in_string = false;
+    let mut escaped = false;
+    let mut line_comment = false;
+    let mut block_comment = false;
+
+    while let Some(ch) = chars.next() {
+        if line_comment {
+            if ch == '\n' {
+                line_comment = false;
+                output.push('\n');
+            }
+            continue;
+        }
+
+        if block_comment {
+            if ch == '*' && chars.peek() == Some(&'/') {
+                let _ = chars.next();
+                block_comment = false;
+            } else if ch == '\n' {
+                output.push('\n');
+            }
+            continue;
+        }
+
+        if in_string {
+            output.push(ch);
+            if escaped {
+                escaped = false;
+            } else if ch == '\\' {
+                escaped = true;
+            } else if ch == '"' {
+                in_string = false;
+            }
+            continue;
+        }
+
+        if ch == '"' {
+            in_string = true;
+            output.push(ch);
+            continue;
+        }
+
+        if ch == '/' && chars.peek() == Some(&'/') {
+            let _ = chars.next();
+            line_comment = true;
+            continue;
+        }
+
+        if ch == '/' && chars.peek() == Some(&'*') {
+            let _ = chars.next();
+            block_comment = true;
+            continue;
+        }
+
+        output.push(ch);
+    }
+
+    output
+}
+
+fn strip_trailing_commas(content: &str) -> std::string::String {
+    let mut output = std::string::String::with_capacity(content.len());
+    let chars: Vec<char> = content.chars().collect();
+    let mut index = 0usize;
+    let mut in_string = false;
+    let mut escaped = false;
+
+    while index < chars.len() {
+        let ch = chars[index];
+        if in_string {
+            output.push(ch);
+            if escaped {
+                escaped = false;
+            } else if ch == '\\' {
+                escaped = true;
+            } else if ch == '"' {
+                in_string = false;
+            }
+            index += 1;
+            continue;
+        }
+
+        if ch == '"' {
+            in_string = true;
+            output.push(ch);
+            index += 1;
+            continue;
+        }
+
+        if ch == ',' {
+            let mut lookahead = index + 1;
+            while lookahead < chars.len() && chars[lookahead].is_whitespace() {
+                lookahead += 1;
+            }
+            if lookahead < chars.len() && matches!(chars[lookahead], '}' | ']') {
+                index += 1;
+                continue;
+            }
+        }
+
+        output.push(ch);
+        index += 1;
+    }
+
+    output
+}
