+++ crates/vize/src/commands/check/tsconfig_inputs/mod.rs
@@ -1,0 +1,138 @@
+//! `tsconfig.json`-driven default input collection for `vize check`.
+//!
+//! When users run `vize check` without explicit paths, we should follow the
+//! project's configured `files` / `include` / `exclude` fields instead of
+//! recursively scanning every TypeScript file under the working directory.
+
+#![allow(clippy::disallowed_macros, clippy::disallowed_types)]
+
+use std::path::{Path, PathBuf};
+
+use vize_carton::FxHashSet;
+
+mod ambient;
+mod collect;
+mod glob;
+mod jsonc;
+mod loader;
+mod matching;
+mod spec;
+
+#[cfg(test)]
+mod tests;
+
+pub(crate) use ambient::collect_ambient_declaration_files;
+pub(crate) use collect::resolve_tsconfig_for_files;
+pub(super) use jsonc::parse_jsonc_value;
+pub(crate) use loader::load_tsconfig_declaration_options;
+pub(super) use loader::{read_extends_entries, resolve_extended_tsconfig};
+pub(crate) use spec::TsconfigDeclarationOptions;
+
+use collect::{
+    collect_supported_files, collect_supported_files_with_options, explicit_hidden_include_roots,
+};
+use glob::{default_exclude_specs, normalize_input_path};
+use loader::{collect_tsconfig_project_paths, load_tsconfig_inputs};
+use matching::is_supported_check_file;
+use spec::{FileCollectionOptions, GlobSpec};
+
+const TARGET_DIR: &str = "target";
+const NODE_MODULES_DIR: &str = "node_modules";
+const VIZE_CACHE_DIR: &str = ".vize";
+
+pub(crate) fn collect_default_check_files(
+    project_root: &Path,
+    tsconfig_path: Option<&Path>,
+) -> Vec<PathBuf> {
+    collect_default_check_files_inner(project_root, tsconfig_path, false)
+}
+
+fn collect_default_check_files_inner(
+    project_root: &Path,
+    tsconfig_path: Option<&Path>,
+    include_hidden_tsconfig_roots: bool,
+) -> Vec<PathBuf> {
+    let Some(tsconfig_path) = tsconfig_path else {
+        return collect_supported_files(project_root, &[], &[]);
+    };
+
+    let mut files = Vec::new();
+    let mut seen = FxHashSet::default();
+    for tsconfig_path in collect_tsconfig_project_paths(tsconfig_path) {
+        collect_default_check_files_for_tsconfig(
+            project_root,
+            &tsconfig_path,
+            include_hidden_tsconfig_roots,
+            &mut files,
+            &mut seen,
+        );
+    }
+
+    files.sort();
+    files
+}
+
+fn collect_default_check_files_for_tsconfig(
+    project_root: &Path,
+    tsconfig_path: &Path,
+    include_hidden_tsconfig_roots: bool,
+    files: &mut Vec<PathBuf>,
+    seen: &mut FxHashSet<PathBuf>,
+) {
+    let spec = load_tsconfig_inputs(tsconfig_path).unwrap_or_default();
+
+    for file in spec.files {
+        let resolved = normalize_input_path(&file.resolve());
+        if resolved.starts_with(project_root)
+            && resolved.is_file()
+            && is_supported_check_file(&resolved)
+            && seen.insert(resolved.clone())
+        {
+            files.push(resolved);
+        }
+    }
+
+    let default_base_dir = tsconfig_path
+        .parent()
+        .map(Path::to_path_buf)
+        .unwrap_or_else(|| project_root.to_path_buf());
+
+    let includes = if !spec.has_includes && !spec.has_files && files.is_empty() {
+        GlobSpec::new(&default_base_dir, "**/*")
+            .into_iter()
+            .collect::<Vec<_>>()
+    } else {
+        spec.includes
+    };
+
+    let excludes = if !spec.has_excludes {
+        default_exclude_specs(&default_base_dir)
+    } else {
+        spec.excludes
+    };
+
+    if !includes.is_empty() {
+        let collected = collect_supported_files(project_root, &includes, &excludes);
+        for path in collected {
+            if seen.insert(path.clone()) {
+                files.push(path);
+            }
+        }
+        if include_hidden_tsconfig_roots {
+            for root in explicit_hidden_include_roots(project_root, &includes) {
+                for path in collect_supported_files_with_options(
+                    &root,
+                    &includes,
+                    &excludes,
+                    FileCollectionOptions {
+                        include_hidden: true,
+                    },
+                ) {
+                    if seen.insert(path.clone()) {
+                        files.push(path);
+                    }
+                }
+            }
+        }
+    }
+}
